backend/pyproject.toml
[tool.poetry]
name = "stori-newsletter-sending-app"
version = "0.1.0"
description = ""
authors = ["Luis Guillermo Rubio <memorubioc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
mysql-connector-python = "^8.0.29"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.18.2"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
